python/Plotly/requirements.txt
dash
plotly
databricks-connect>=15.1.0